refactor: 拆分 jjc_ranking 为 api_client/cache_repo
